Reorganize classes.
