+++ src/main/java/info/coremodding/craftenchants/event/PreInitializer.java
@@ -5,14 +5,14 @@
  */
 package info.coremodding.craftenchants.event;
 
-import info.coremodding.craftenchants.item.CEItems;
+import info.coremodding.craftenchants.item.ItemsCE;
 
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
 
 public final class PreInitializer {
     
     public static void handle(FMLPreInitializationEvent event) {
-	CEItems.initialize();
+	ItemsCE.initialize();
     }
     
 }
