+++ src/main/java/info/coremodding/craftenchants/item/crafting/CEItemRecipe.java
@@ -3,7 +3,7 @@
  * http://www.gnu.org/licenses/gpl.html
  * http://www.gnu.org/licenses/lgpl.html
  */
-package info.coremodding.craftenchants.item;
+package info.coremodding.craftenchants.item.crafting;
 
 import net.minecraft.init.Blocks;
 import net.minecraft.init.Items;
